# perf(core): check files before interacting with them