--- a/packages/babel-core/src/gensync-utils/fs.ts
+++ b/packages/babel-core/src/gensync-utils/fs.ts
@@ -10,14 +10,8 @@ export const readFile = gensync<(filepath: string, encoding: "utf8") => string>(
 
 export const exists = gensync<(filepath: string) => boolean>({
   sync(path) {
-    try {
-      fs.accessSync(path);
-      return true;
-    } catch {
-      return false;
-    }
+    return fs.existsSync(path);
   },
-  errback: (path, cb) => fs.access(path, undefined, err => cb(null, !err)),
 });
 
 export const stat = gensync<typeof fs.statSync>({