--- a/packages/babel-core/src/gensync-utils/fs.ts
+++ b/packages/babel-core/src/gensync-utils/fs.ts
@@ -8,12 +8,6 @@ export const readFile = gensync<(filepath: string, encoding: "utf8") => string>(
   },
 );
 
-export const exists = gensync<(filepath: string) => boolean>({
-  sync(path) {
-    return fs.existsSync(path);
-  },
-});
-
 export const stat = gensync<typeof fs.statSync>({
   sync: fs.statSync,
   errback: fs.stat,