@nicolo-ribaudo 2021-04-02
I think in the past we already used `existsSync`, but then switched to `accessSync` because `fs.exists` (the async version) is deprecated and we got confused.