@nicolo-ribaudo 2021-04-02
> This is technically a semantic change, if you previously had files that you couldn't read, you will now see an error when trying to read them (after this function), instead of them being silently ignored. I think this is better behaviour. 

That's a good change imo. If someone is passing a file to Babel and Babel cannot read it, it's better to throw an explicit error than to silently not do what the user is asking Babel to do.